refactor: harmonize filenames to strictly match dominant exported symbols per AGENTS.md
