+++ src/client/mod.ts
@@ -6,4 +6,4 @@ export * from "./search-index/mod.ts";
 export * from "./providers/rdfjs/mod.ts";
 export * from "./providers/comunica/mod.ts";
 
-export * from "./providers/libsql/libsql-provider.ts";
+export * from "./providers/libsql/provide-libsql.ts";
