+++ src/client/providers/libsql/mod.ts
@@ -1,4 +1,4 @@
 export * from "./libsql-search-index.ts";
-export * from "./libsql-provider.ts";
-export * from "./libsql-quad-hydrator.ts";
-export * from "./libsql-quad-synchronizer.ts";
+export * from "./provide-libsql.ts";
+export * from "./hydrate-store-from-libsql.ts";
+export * from "./sync-libsql.ts";
