+++ src/client/quad-store/create-indexed-store.test.ts
@@ -1,6 +1,6 @@
 import { assertEquals } from "@std/assert";
 import { DataFactory, Store } from "n3";
-import { createIndexedStore } from "./indexed-store.ts";
+import { createIndexedStore } from "./create-indexed-store.ts";
 import { QueryEngine } from "@comunica/query-sparql-rdfjs-lite";
 import { executeSparql } from "#/client/providers/comunica/mod.ts";
