+++ src/client/quad-store/create-indexed-store.ts
