+++ src/client/quad-store/hash-quad.ts
