+++ src/client/quad-store/mod.ts
@@ -1,3 +1,3 @@
 export * from "./interface.ts";
-export * from "./quad-store.ts";
+export * from "./rdfjs-quad-store.ts";
 export * from "./patch.ts";
