+++ src/client/quad-store/rdfjs-quad-store.ts
